add-empty-branch: don't make git compare abyss with our normal remote
If we add abyss as a remote and do "git fetch", it tries to find a
common ancestor, which pretty much means telling the server all of the
objects we know about.  It's slow as hell and *also* leaks data out to
the world.

Instead, clone abyss into a temporary repo, then add *that* as the
remote, and then clean it all up when we're done.

Signed-off-by: Peter Jones <pjones@redhat.com>